docs: move shortcodes to their own plugin
